renderGlows: type alpha as u8 to store colorScale via direct stb
Target holds the 0xff/computed alpha in a saved register and stores it to
colorScale (u8 global) with a bare stb; the int local forced a clrlwi
truncation at the store. Typing alpha u8 matches the direct byte store.
renderGlows fuzzy 94.72->94.75, no sibling regression, ninja EXIT=0.

+++ src/main/track_dolphin.c
@@ -3494,7 +3494,7 @@ void renderGlows(void)
     f32 sunMtx[12];
     f32 dir[3];
     f32 cam[3];
-    int alpha;
+    u8 alpha;
     u8 sky;
     f32 sunDot;
